+++ app/platforms/fate/character.rb
@@ -13,6 +13,7 @@ module Fate
     attribute :custom_stress, array: true, default: []
     attribute :selected_stress, array: true, default: {} # physical 3, mental: 3
     attribute :consequences, array: true, default: {} # mild,moderate,severe, physical/mental
+    attribute :stunts, array: true, default: [] # [{ id: 1, title: '', description: '', skill: nil }]
   end
 
   class Character < Character