+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - vitals management for Fate characters
+- stunts management for Fate characters
 
 ## [0.4.6] - 2026-02-09
 ### Added